#9: WIP Create workflow code-quality and manage automated testing of basic cases with tox
