+++ tests/config/example_1.yaml
@@ -1,0 +1,2 @@
+example_id: 1
+is_fwmp: False
